Rope Climbable implementation

+++ UE_DungeonCompany/Source/UE_DungeonCompany/Public/WorldActors/Ladder.h
@@ -44,7 +44,7 @@ private:
 	FVector2D EasyInteractArea = FVector2D(5, 20);
 
 public:
-	virtual FVector GetLocationAtZ(float Z) const override;
+	virtual FVector GetLocationAtZ(double Z) const override;
 	virtual FVector GetUpperEndLocation() const override;
 
 	/*This doesn't set the height in runtime. Set this in the constructor or before calling the construction script of the Ladder class.*/
